refactor: 유틸 함수 이동 및 NavigationStrategy 파일 삭제
